use double for font sizing
git-svn-id: https://svn.r-project.org/R/trunk@44645 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/Graphics.h
@@ -109,7 +109,7 @@ typedef struct {
     int	pch;		/* Plotting character */
     /* Note that ps is never changed, so always the same as dev->startps.
        However, the ps in the graphics context is changed */
-    int ps;		/* Text & symbol pointsize */
+    double ps;		/* Text & symbol pointsize */
     int	smo;		/* Curve smoothness */
     double srt;		/* String Rotation */
     double tck;		/* Tick size as in S */
